build: refactoring moving code to separate file outside of init
