+++ .flake8
@@ -1,0 +1,4 @@
+[flake8]
+ignore = E501, W605, F901, F401
+max-line-length = 120
+max-complexity = 10
